LU-8573 ptlrpc: cleanup bulks for ptlrpc_register_bulk failure
It is assumed that if ptlrpc_register_bulk() failed, then there
should be no bulk registered. But for the case of transferring
multiple bulks, it may be some bulks registered, but the others
not. Under such case, we need to cleanup the request environment
via ptlrpc_unregister_bulk() explicitly to avoid misguiding when
re-try the request some time later.

On the other hand, for replay case, to be safe, we cleanup the
request environment via ptlrpc_unregister_bulk() before sending
out the replay request.

More sanity check about the bulk MDs in the ptlrpc logic.

Signed-off-by: Fan Yong <fan.yong@intel.com>
Change-Id: Ibc1063fdd614dcaf8671c5a3c6f820f8ade417d9