LU-7481 utils: commit device label
When changing the device label for the first mount, tune2fs reads
the buffer containing the device super_block and writes it back
after some modifications, if there are no super_block related
operations in the corresponding ldiskfs filesystem, the buffer
won't be written back to disk along with the journal.

This patch calls fsync on the device to sync the change to disk.

Change-Id: Ie7a78e28c3592f2b45e71613b3f64e53ba1430e4
Signed-off-by: Hongchao Zhang <hongchao.zhang@intel.com>