LU-6688 tests: use proper nodes for NRS test
Several of the NRS test for sanityn are reporting
error: set_param: ost/OSS/*/nrs_policies: Found no match.
The reason for this is that the test are attempting
to configure NRS oss settings on the MDS servers.
Those oss settings don't exist on MDS servers.
Change the test to alter the oss NRS settings on
the OSS servers instead.

Change-Id: I83600165fc1b9f0d9c6ee0d093f54604c46328b9
Test-Parameters: trivial testlist=sanityn
Signed-off-by: James Simmons <uja.ornl@yahoo.com>