LU-8569 linkea: linkEA size limitation
Under DNE mode, if we do not restrict the linkEA size, and if there
are too much cross-MDTs hard links to the same object, then it will
casue the llog overflow. On the other hand, too much linkEA entries
in the linkEA will serious affect the linkEA performance because we
only support to locate linkEA entry consecutively.

So we need to restrict the linkEA size. Currently, it is 4096 bytes,
that is independent from the backend. If too much hard links caused
the linkEA overflowed, we will add overflow timestamp in the linkEA
header. Such overflow timestamp has some functionalities:

1. It will prevert the object being migrated to other MDT, because
   some name entries may be not in the linkEA, so we cannot update
   these name entries for the migration.

2. It will tell the namespace LFSCK that the 'nlink' attribute may
   be more trustable than the linkEA, then avoid misguiding the
   namespace LFSCK to repair 'nlink' attribute based on linkEA.

If the linkEA is marked as overflow, and if some hard links are
removed, then we still cannot clear the overflow timestamp, because
the missed hard links entries are still not in the linkEA. It is the
namespace LFSCK duty to add the missed hard links entries into the
linkEA (if possible) and clear the overflow timestamp if all the
hard links entries are in the linkEA. But it is possible that there
are still new hard links added during the namespace LFSCK scanning.
So the namespace LFSCK needs to compare the linkEA overflow timestamp
with the namespace LFSCK latest reset time in the 2nd-stage scanning.
If the latter one is newer, then means all the hard links entries have
been in the linkEA, so the linkEA overflow timestamp can be cleared.

The old interfaces between MDD/OUT and LFSCK via lfsck_in_notify() for
the linkEA overflow event are obsoleted and removed.

Test-Parameters: mdtcount=2 mdscount=2 mdsfilesystemtype=zfs mdtfilesystemtype=zfs ostfilesystemtype=zfs envdefinitions=ONLY=29d testlist=sanity-lfsck
Test-Parameters: mdtcount=2 mdscount=2 mdsfilesystemtype=zfs mdtfilesystemtype=zfs ostfilesystemtype=zfs envdefinitions=ONLY=409 testlist=sanity
Signed-off-by: Fan Yong <fan.yong@intel.com>
Change-Id: I2d6c2be04305c1d7e3af160d8b80e73b66a36483