LU-8455 subject: Pacemaker script for Lustre and ZFS
A new script to be used in Pacemaker to manage ZFS pools and Lustre
targets.
This RA is able to manage (import/export) ZFS pools and Lustre Target
(mount/umount).
pcs resource create <Resource Name> ocf:heartbeat:LustreZFS
pool="<ZFS Pool Name>"
volume="<ZFS Volume Name>"
mountpoint="<Mount Point"
OCF_CHECK_LEVEL=10

where:
pool is the pool name of the ZFS resource created in advance volume
is the volume name created on the ZFS pool during the Lustre format
(mkfs.lustre).
mount point is the mount point created in advance on both the Lustre
servers OCF_CHECK_LEVEL is optional and enable an extra monitor on
the status of the pool.

This script should be located in /usr/lib/ocf/resource.d/heartbeat/
of both the Lustre servers with permission 755.

The script provides protection from double imports of the pools. In
order to activate this functionality is important to configure the
hostid protection in ZFS using the genhostid command.

At the moment the script is designed to manage 2-way cluster only.

Test-Parameters: trivial
Signed-off-by: Gabriele Paciucci <gabriele.paciucci@intel.com>
Change-Id: I068aad7ceb8f7cd987f54c0a55df202e936fb5cf