LU-8507 lnet: Enable setting per NI peer_credits
The code to allow peer_credits to be set per NI was originally
"left inactive" because there were concerns about peer_credits
interfering with the ability for IB nodes to connect to each
other when peer_credits are not the same (peer_credits controls
the queue depth for IB).
With LU-3322, the values do not have to match so it is now safe
to enable this code so peer_credits can be set per NI.

This patch enables existing code for setting per NI peer_credits.

Test-Parameters: trivial
Signed-off-by: Doug Oucharek <doug.s.oucharek@intel.com>
Change-Id: I28ede7a139c43ca9a3d1b22255d3358694057918