LU-8900 snapshot: operate write barrier on MDT
Currently, the Lustre barrier is implemented as write barrier on
all MDTs. For each MDT in the system, when it starts, it registers
a barrier instance that will start a barrier engine. The barrier
engine only serves this barrier instance. So the barrier is per-system
based, not per-node based. Means that even if multiple Lustre systems
share the same MGS node or multiple MDTs on the same MDS node, we can
control the barrier on specified system but NOT affect others.

The barrier_handler() converts the incoming barrier request as local
format, and adds the converted barrier request to related barrier
instance request queue, then wakes up related barrier engine to handle
such request.

The barrier engine fetches barrier request from the barrier instance
request queue. According to the request's status, it will freeze the
barrier or thaw the barrier, then notify the MGS about the result of
handling the barrier request.

On the other hand, if the barrier is not thawed within its life cycle,
means expired, then the local barrier_engine will thaw the barrier
automatically to avoid the MDT to be frozen for ever.

Signed-off-by: Fan Yong <fan.yong@intel.com>
Change-Id: I8d209e98d175eacdadd25c385ffc1c3e4451527a