LU-5964 tests: allow createmany to replace unlinkmany
The functionality of unlinkmany is really just a subset of createmany.
Allow createmany to replace unlinkmany by checking argv[0] and setting
the "do_unlink" flag in this case and disallowing the create options.

That allows removal of the code duplication that is unlinkmany.c.

Fix replay-single.sh test_11b to not pass "-o" as an argument.  This
was a copy-paste error from the "createmany -o" line above, but fails:

    CMD: unlinkmany -o /mnt/lustre2/f11b.replay-vbr-2- 100
    CMD: createmany -o /mnt/lustre/f11b.replay-vbr-1- 100
    unlink(-o0) error: No such file or directory
    total: 0 unlinks in 0 seconds: -nan unlinks/second

Test-Parameters: trivial testlist=replay-vbr
Signed-off-by: Andreas Dilger <andreas.dilger@intel.com>
Change-Id: I8f44b33a5374121dc25538ac2029a8050af777c5