LU-7734 lnet: match interface ip address when parsing ip2nets
The "ip2nets" module option and "ip2net" lnetctl option make it
possible to specify which LNet network must be instantiated by
giving rules that match IP addresses. The first match of an IP
address rule against an interface on the node is used.  The code
then constructs a "networks" line that is parsed as such to
create the LNet network.

However, when instantiating the LNet network the code uses the
first interface regardless of whether it matches the address
pattern. Change the code to track the matches by constructing
a network line that includes an interface list. For Multi-Rail
nodes it is expected that multiple interfaces exist on the same
network, and these are all added to the list.

If the ip2nets network specification already contains an explicit
interface list, then that list will not be changed.

Setting the lnet_old_ip2nets_parser module parameter restores the
old behavior.

As an illustration of the problem, take a host with the following
network interfaces defined, and note that interfaces 3, 10, and 11
are in the same subnet.

 1: lo:           127.0.0.1/8
 2: eno16777736:  192.168.138.17/24
 3: eno33554984:  192.168.131.17/24
 4: eno50332208:  192.168.132.17/24
 5: eno67109432:  192.168.133.17/24
 6: eno83886656:  192.168.134.17/24
 7: eno100663880: 192.168.135.17/24
 8: eno117441104: 192.168.136.17/24
 9: eno134218328: 192.168.137.17/24
10: eno150995552: 192.168.131.27/24
11: eno167772776: 192.168.131.37/24

With the old code the first non-loopback interface is configured,
despite it being on a different network than specified by the pattern.

[root@lustre-client ~]# lnetctl net add --ip2net "tcp1 192.168.131.*"
[root@lustre-client ~]# lnetctl net show
net:
    - net type: lo
      local NI(s):
        - nid: 0@lo
          status: up
    - net type: tcp1
      local NI(s):
        - nid: 192.168.138.17@tcp1
          status: up

With the new code the three interfaces whose IP address matches the
pattern have been configured.

[root@lustre-client ~]# lnetctl net add --ip2net "tcp1 192.168.131.*"
[root@lustre-client ~]# lnetctl net show
net:
    - net type: lo
      local NI(s):
        - nid: 0@lo
          status: up
    - net type: tcp1
      local NI(s):
        - nid: 192.168.131.17@tcp1
          status: up
          interfaces:
              0: eno33554984
        - nid: 192.168.131.27@tcp1
          status: up
          interfaces:
              0: eno150995552
        - nid: 192.168.131.37@tcp1
          status: up
          interfaces:
              0: eno167772776

Test-Parameters: trivial
Signed-off-by: Olaf Weber <olaf@sgi.com>
Change-Id: Id7b756512c07d898c1075b83b67f9973c24fd312