LU-7734 lnet: add lnet_max_interfaces tunable
Add an lnet_max_interfaces tunable value, that describes the maximum
number of interfaces per node. This tunable is primarily useful for
sanity checks prior to allocating memory.

Add LNET_MIN_INTERFACES, value 16, as the minimum value.

Add LNET_MAX_INTERFACES_DEFAULT, value 200, as the default value. This
value was chosen to ensure that the size of an LNet ping message with
any associated LND overhead would fit in 4096 bytes.

(The LNET_MAX_INTERFACES name was not reused to allow for the early
detection of issues when merging code that uses it.)

Test-Parameters: trivial
Signed-off-by: Olaf Weber <olaf@sgi.com>
Change-Id: I1fd88e9e37d21614bb0af8a4fe14d9ae5ae62c1e