LU-7734 lnet: resend LNet messages on failure
*** experimental code ***

When lnet_finalize() is called with a non-zero status, attempt to
resend the message if the deadline has not yet passed.

The deadline is set to lnet_msg_timeout milliseconds after the call
to lnet_prep_send() that prepares the message for sending. If the
deadline expires before the message is sent for the first time, it
will still be sent: the deadline only applies to the decision to
retry.

The default value of lnet_msg_timeout is 1000ms.

Resending a message requires selectively returning the TX credits
before calling lnet_send() again. lnet_msg_uncommit_tx() does this.

The source and router NIDs that used to be lnet_send() parameters
are stored in the lnet_msg, so that we can resend with the same
values.

The actual resending is done from lnet_complete_msg_locked().

Test-Parameters: trivial
Signed-off-by: Olaf Weber <olaf@sgi.com>
Change-Id: I1bd73b513ac76da7e4c4dc737f027f464d774dfb