LU-8633 osc: check if vmpage is in use in osc_lru_shrink
lru_page_busy should always check if the vmpage is in use,
otherwise osc_lru_shrink could evict a page which is read in
by another thread handling a page fault and waiting on it
to become up to date.
Evicting such a page would cause a SIGBUS delivered to the
thread handling the page fault.

Signed-off-by: Li Dongyang <dongyang.li@anu.edu.au>
Change-Id: I8a9ddcc3a494d0bd8f4c2fe916b3e4efe006d5c6