LU-8674 obd: Add a general policy engine
This patch adds a general policy engine. And also HSM support
is added for it.

Currently, only one archive ID is supported. And the archive ID
can be set by writing the ID to following proc:

/proc/fs/lustre/mdt/*/hsm_policy_archive_id

And an rule could be writting to following proc:

/proc/fs/lustre/mdt/*/hsm_policy_rule

The rule should be in the form of Polish notation. A rule that will
trigger HSM archive action if the modify timestamp of the file is 1
minute ealier than the system time looks like this:

& - >= mtime - sys_time 60 1 hsma_bit_archive

By writing the FID to the following proc, the value of the rule on
the file could be printed by reading the proc of hsm_policy_rule.

/proc/fs/lustre/mdt/*/hsm_policy_dump_obj

Signed-off-by: Li Xi <lixi@ddn.com>
Change-Id: I5b523e1bc7bf8372de5da5ab2a7cb72342d9608e