LU-8633 llite: do not clear uptodate bit in page delete
Otherwise, if there exists race between page fault and truncate, it
will cause the page fault routine to return EIO error.

In filemap_fault() {
    page_not_uptodate:
	...
        ClearPageError(page);
        error = mapping->a_ops->readpage(file, page);
        if (!error) {
                wait_on_page_locked(page);
                if (!PageUptodate(page))
                        error = -EIO;
	}
	...
}

However, I tend to think this is a defect in the implementation,
because it assumes PageUptodate shouldn't be cleared but file read
routine doesn't make the same assumption.

Signed-off-by: Jinshan Xiong <jinshan.xiong@intel.com>
Change-Id: Ic4a919607a6121098e41eaf56b8ce3200f778ecf