LU-8656 test: test IS_NOSEC and i_mutex usage
We must use the i_mutex to protect permission changes,
which means we need to take it when we write to a file with
the setuid or setgid bit set (as this removes those bits).

We use IS_NOSEC to check for this case.  This test verifies
that check is working correctly.

NOTE TO BE REMOVED LATER:
Currently, that check is wrong, and is causing us to always
take the i_mutex.  This test is partly to see if that's
true on all supported platforms.

Signed-off-by: Patrick Farrell <paf@cray.com>
Change-Id: If481366283ceaeb247e50cc4d536fb10e9c329ee