LU-8653 lod: use stripe_count instead of stripe_nr
Replace the use of stripenr in the code with stripe_count so that
this is consistent with this field name in the rest of the code.
Move ldo_stripe_count and ldo_stripes_allocated out of the lod_object
union instead of only a comment that the union fields are shared.

Introduce LOV_PATTERN_NONE instead of using "0" around the code to
indicate no layout has been specified.

Introduce LOV_PATTERN_DEFAULT to indicate that the entire layout is
unset, instead of using "0xffffffff" explicitly in the code.

Signed-off-by: Andreas Dilger <andreas.dilger@intel.com>
Change-Id: I6056aebc1a381b09c1a436fb4a7986a51f5a0c30