LU-1573 recovery: Avoid data corruption for DIO during FOFB
recovery: DIO requests should not be sent on replay Q

When there is a userland app doing DIO and OBD failover,
obd->obd_no_transno is set to 1 & last_committed on server
is not sent to the client. Thus client is not sure if the
req s _committed_ to disk or not. So it removes the req
from resend queue and adds it to replay queue.

Now trans_no > last_committed, thus after reconnect, as a
part of recovery process, request is replayed. Userland app,
refills the DIO buffer with different data, thus invalid data
is committed resulting in corruption.

This change avoids the client replay by sending across the
last_committed to client and also avoids the replay and the
corruption.

tests: test data integrity during OSS failover

replay-single.sh test_70f checks to see if the scenario in MRP-542
is reproducible.

Seagate-Bug-Id: MRP-542, MRP-2418

Signed-off-by: Alexey Lyashkov <alexey.lyashkov@seagate.com>
Signed-off-by: Parinay Kondekar <parinay.kondekar@seagate.com>
Signed-off-by: Andrew Perepechko <andrew.perepechko@seagate.com>
Signed-off-by: Emoly Liu <emoly.liu@intel.com>
Change-Id: Ia30783c99e6c16a0c7ab70841eb98ed75dba1de9
Reviewed-by: Alexander Zarochentsev <alexander.zarochentsev@seagate.com>
Reviewed-by: Vitaly Fertman <vitaly.fertman@seagate.com>